Create README.md
It is an online game where you create a virtual team of real cricket players and score points depending on how your chosen players perform in real life matches. To win a tournament, you must try and get the maximum points and the No. 1 rank amongst other participants. Here's how a Fantasy Cricket game may look like.

In this cricket game, we can create new team or open an existing team. When we create a new team, we have to enter team name. Then we are given 1000 points to select total 11 players from the team. Select the players carefully from list 1 till your points exhaust. We can select total 4 batsman, 3 bowlers, 3 all-rounder and 1 wicket-keeper. We can then save the team. While saving the team, it checks whether the condition is fulfilled. The team is saved in the database sqlite3. We can also open existing team from the database. When we clicked on evaluate, a new dialog box is opened were team is being evaluated for their performance. In evaluation process, we have to select particular team to be evaluated from the combobox1. When we select the team, the players table is populated. When we click calculate score, the points table is populated according to their stats in the database. The points are calculated according to the rules. Then the total points are being calculated. Team whose points is maximum wins the game.